LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/html/browsers/origin/origin-of-data-document-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.



PASS The origin of a 'data:' document in a frame is opaque.
